started 3 stage pipeline - lots of general fixes & TODO: debugging and fixing processes

+++ src/defns.h
@@ -4,9 +4,10 @@
 /*
  * Type Definitions for types used across project
  */
+#include <stdint.h>
 
 typedef unsigned char byte_t;
-typedef unsigned long int word_t;
+typedef uint32_t word_t;
 typedef unsigned char nibble_t;
 typedef unsigned char bit_t;
